% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elisa4ll.R
\name{elisa4ll}
\alias{elisa4ll}
\title{Calculate the concentration of a substrate from a known OPTICAL DENSITY}
\usage{
elisa4ll(OD, conc, OD_sample)
}
\arguments{
\item{OD}{Optical density of the STANDARD DILUTION [numeric vector]}

\item{conc}{KNOWN concentrations of the standard dilution [numeric vector]}

\item{OD_sample}{Optical densities of samples [named vector]}
}
\value{
A data frame with OD, log fitted values and calculated concentrations
 of the samples and a plot with a standard curve standard and blue samples
}
\description{
This function will give a calculated concentration of a subsance in 96 well 
plate given the optical density from a plate reader.
}
\details{
This function is useful when working with old plate readers. It uses 
the FOUR PARAMETER CURVE to fit the model to your standards.
}
\author{
Wojciech Francuzik
}
