% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateMap.R
\name{plateMap}
\alias{plateMap}
\title{plate map function to make a quick plate map out of given conditions}
\usage{
plateMap(rows_var_name = "medium", cols_var_name = "strain", rows, cols)
}
\description{
plate map function to make a quick plate map out of given conditions
}

