% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlateMap.R
\name{plotPlateMap}
\alias{plotPlateMap}
\title{Ploting the graphical map of a designed template for experimental plan.}
\usage{
plotPlateMap(platemap)
}
\arguments{
\item{platemap}{is an dataframe with following variables: Column, Row, medium, strain}
}
\value{
a ggplot
}
\description{
This function will let you make a map of a 96 well plate for planning your experiments
}
\details{
this was modified from the work by Brian D. Connelly
}
\author{
Wojciech Francuzik
}

