% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_OD_broth.R
\name{prepare_OD_broth}
\alias{prepare_OD_broth}
\title{Calculate how to get a desired optical density}
\usage{
prepare_OD_broth(is, desired, is_vol)
}
\arguments{
\item{is}{hive your measured OD here}

\item{desired}{the desired OD}

\item{is_vol}{the volum in which you curretly have the measured OD}
}
\value{
a character string telling you what to do
}
\description{
This function will give a prescription for the correct OD of sample, using the
known OD of a bacterial broth, the volume of this brot which willl be the same a the newly aquired sample,
all values in microliters, for rho please use "grams/1000000" to get the correct value.
}
\details{
this function is helpfull while working in the lab
}
\author{
Wojciech Francuzik
}
