% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.molarity.R
\name{solution.molarity}
\alias{solution.molarity}
\title{This function calculates the molar concentration of something
 please provide:
 1. volume of the used substance (microliters) of interest that was dissolved in
 2. volume of the medium in which the substance was dissolved
 3. molar weight of the substance of interest
 4. Percentage of the substance of interest expressed as a fraction (0.37 for a 37% acid)
 5. substance density - rho expressed in grams per 1000000 microliteres (1 liter)
You will get the value of molar concentration of the calculated medium expressed in mols per liter}
\usage{
solution.molarity(substance_vol, medium_vol, molar_weight, solution_percent,
  substance_rho)
}
\description{
This function calculates the molar concentration of something
 please provide:
 1. volume of the used substance (microliters) of interest that was dissolved in
 2. volume of the medium in which the substance was dissolved
 3. molar weight of the substance of interest
 4. Percentage of the substance of interest expressed as a fraction (0.37 for a 37% acid)
 5. substance density - rho expressed in grams per 1000000 microliteres (1 liter)
You will get the value of molar concentration of the calculated medium expressed in mols per liter
}
